+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - CLI command to emulate stop button being pressed.
+- CLI command to print active alarms.
 
 
 ## [0.6.1] - 2022-04-27