+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CLI command to emulate stop button being pressed.
 - CLI command to print active alarms.
 - CLI now also sends Serial_change_character when alarm configuration changes.
+- CLI command to print all status information at once (`status`). Useful as a
+  response to `Serial_change_character` to know what changed.
 
 
 ## [0.6.1] - 2022-04-27