+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [0.7.0] - 2023-11-05
 ### Added
 - CLI command to emulate stop button being pressed.
 - CLI command to print active alarms.